requirements.txt
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
gTTS==2.0.3
pydub==0.23.1
soundfile==0.10.2
astroid==2.2.5
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
imgkit==1.0.2
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
moviepy==1.0.0
numpy==1.16.3
pandas==0.24.2
Pillow==6.2.0
praw==6.2.0
prawcore==1.0.1
oauth2client==4.1.3
proglog==0.1.9
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rope==0.14.0
six==1.12.0
tqdm==4.32.1
typed-ast==1.3.2
update-checker==0.16
urllib3==1.25.2
websocket-client==0.56.0
wrapt==1.11.1
tinydb
